+++ apps/frontend-manage/src/pages/sessions/[id]/lecturer.tsx
@@ -2,10 +2,8 @@ import { useQuery } from '@apollo/client'
 import { faThumbsUp } from '@fortawesome/free-regular-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import {
-  AggregatedConfusionFeedbacks,
   Feedback,
   GetPinnedFeedbacksDocument,
-  Session,
 } from '@klicker-uzh/graphql/dist/ops'
 import dayjs from 'dayjs'
 import Head from 'next/head'
@@ -36,13 +34,12 @@ function Feedbacks() {
   const aggregateConfusion = useMemo(() => {
     if (
       data?.pinnedFeedbacks &&
-      data?.pinnedFeedbacks.confusionFeedbacks &&
-      data?.pinnedFeedbacks.confusionFeedbacks[0] &&
-      data?.pinnedFeedbacks.confusionFeedbacks[0].numberOfParticipants > 0
+      data?.pinnedFeedbacks.confusionSummary &&
+      data?.pinnedFeedbacks.confusionSummary.numberOfParticipants > 0
     ) {
       return Math.max(
-        Math.abs(data?.pinnedFeedbacks.confusionFeedbacks[0].speed),
-        Math.abs(data?.pinnedFeedbacks.confusionFeedbacks[0].difficulty)
+        Math.abs(data?.pinnedFeedbacks.confusionSummary.speed),
+        Math.abs(data?.pinnedFeedbacks.confusionSummary.difficulty)
       )
     }
     return 0
@@ -71,13 +68,11 @@ function Feedbacks() {
   const {
     isLiveQAEnabled,
     isConfusionFeedbackEnabled,
-    confusionFeedbacks,
+    confusionSummary,
     feedbacks,
-  } = data?.pinnedFeedbacks as Session & { feedbacks: Feedback[] } & {
-    confusionFeedbacks: AggregatedConfusionFeedbacks[]
-  }
+  } = data?.pinnedFeedbacks
 
-  if (!isLiveQAEnabled) {
+  if (!isLiveQAEnabled && !isConfusionFeedbackEnabled) {
     return <div className="p-4">Publikumsinteraktion ist nicht aktiviert.</div>
   }
 
@@ -96,8 +91,9 @@ function Feedbacks() {
         <title>Feedbacks</title>
       </Head>
 
-      <div className="flex-1">
-        {/* // TODO: readd dropdown to allow lecturer to sort feedbacks according to preferences
+      {isLiveQAEnabled && (
+        <div className="flex-1">
+          {/* // TODO: readd dropdown to allow lecturer to sort feedbacks according to preferences
         <Dropdown
           selection
           options={[
@@ -108,36 +104,38 @@ function Feedbacks() {
           onChange={(_, { value }) => setSortBy(value as string)}
         /> */}
 
-        {feedbacks?.length === 0 && (
-          <div className="mt-4 flex items-center border border-solid bg-primary-10% border-primary text-2xl p-4">
-            No feedbacks received or pinned yet...
-          </div>
-        )}
+          {feedbacks?.length === 0 && (
+            <div className="mt-4 flex items-center border border-solid bg-primary-10% border-primary text-2xl p-4">
+              No feedbacks received or pinned yet...
+            </div>
+          )}
 
-        {feedbacks.map(({ id, content, createdAt, votes }: Feedback) => (
-          <div
-            className="mt-4 flex items-center border border-solid bg-primary-10% border-primary"
-            key={id}
-          >
-            <div className="flex-1 p-4">
-              <p className="mb-0 text-2xl">{content}</p>
-              <div className="flex flex-row mt-2 text-lg text-gray-500">
-                <div>{dayjs(createdAt).format('DD.MM.YYYY HH:mm')}</div>
+          {feedbacks.map(({ id, content, createdAt, votes }: Feedback) => (
+            <div
+              className="mt-4 flex items-center border border-solid bg-primary-10% border-primary"
+              key={id}
+            >
+              <div className="flex-1 p-4">
+                <p className="mb-0 text-2xl">{content}</p>
+                <div className="flex flex-row mt-2 text-lg text-gray-500">
+                  <div>{dayjs(createdAt).format('DD.MM.YYYY HH:mm')}</div>
+                </div>
+              </div>
+              <div className="flex flex-col justify-between flex-initial pr-4">
+                <div className="text-3xl text-gray-500">
+                  {votes}{' '}
+                  <FontAwesomeIcon icon={faThumbsUp} className="mr-0.5" />
+                </div>
               </div>
             </div>
-            <div className="flex flex-col justify-between flex-initial pr-4">
-              <div className="text-3xl text-gray-500">
-                {votes} <FontAwesomeIcon icon={faThumbsUp} className="mr-0.5" />
-              </div>
-            </div>
-          </div>
-        ))}
-      </div>
+          ))}
+        </div>
+      )}
 
       {isConfusionFeedbackEnabled && (
         <div className="flex-initial">
           <div className="flex-initial p-4 w-[300px] bg-primary-bg rounded shadow print:hidden border-primary border-solid border">
-            <ConfusionCharts confusionValues={confusionFeedbacks[0]} />
+            <ConfusionCharts confusionValues={confusionSummary} />
           </div>
         </div>
       )}