+++ apps/frontend-pwa/src/pages/course/[courseId]/element/[id].tsx
@@ -6,13 +6,16 @@ import {
 import { faCheck, faRepeat, faShuffle } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import {
+  ChoicesQuestionData,
+  FreeTextQuestionData,
   GetLearningElementDocument,
+  NumericalQuestionData,
+  QuestionType,
   ResponseToQuestionInstanceDocument,
 } from '@klicker-uzh/graphql/dist/ops'
 import Markdown from '@klicker-uzh/markdown'
 import { addApolloState, initializeApollo } from '@lib/apollo'
 import { getParticipantToken } from '@lib/token'
-import { QuestionType } from '@type/app'
 import { Button, H3, Progress } from '@uzh-bf/design-system'
 import dayjs from 'dayjs'
 import { GetServerSideProps } from 'next'
@@ -37,7 +40,7 @@ interface Props {
 // TODO: leaderboard and points screen after all questions have been completed?
 // TODO: different question types (FREE and RANGE)
 function LearningElement({ courseId, id }: Props) {
-  const [response, setResponse] = useState<number[] | string | null>(null)
+  const [response, setResponse] = useState<{} | number[] | string | null>(null)
   const [currentIx, setCurrentIx] = useState(-1)
 
   const router = useRouter()
@@ -52,11 +55,12 @@ function LearningElement({ courseId, id }: Props) {
   useEffect(() => {
     if (questionData?.type) {
       if (
-        questionData.type === QuestionType.SC ||
-        questionData.type === QuestionType.MC ||
-        questionData.type === QuestionType.KPRIM
+        questionData.type === QuestionType.Sc ||
+        questionData.type === QuestionType.Mc
       ) {
         setResponse([])
+      } else if (questionData.type === QuestionType.Kprim) {
+        setResponse({})
       } else {
         setResponse('')
       }
@@ -89,14 +93,7 @@ function LearningElement({ courseId, id }: Props) {
       variables: {
         courseId: router.query.courseId as string,
         id: currentInstance?.id as number,
-        response:
-          questionData?.type === QuestionType.SC ||
-          questionData?.type === QuestionType.MC ||
-          questionData?.type === QuestionType.KPRIM
-            ? {
-                choices: response as number[],
-              }
-            : { value: response as string },
+        response: formatResponse(questionData, response),
       },
     })
   }
@@ -338,6 +335,30 @@ function LearningElement({ courseId, id }: Props) {
       />
     </Layout>
   )
+}
+
+const formatResponse = (
+  questionData:
+    | ChoicesQuestionData
+    | FreeTextQuestionData
+    | NumericalQuestionData
+    | undefined,
+  response: {} | number[] | string
+) => {
+  if (
+    questionData?.type === QuestionType.Sc ||
+    questionData?.type === QuestionType.Mc
+  ) {
+    return { choices: response as number[] }
+  } else if (questionData?.type === QuestionType.Kprim) {
+    return {
+      choices: Object.keys(response).flatMap<number[]>((key) =>
+        response[key] === true ? [parseInt(key)] : []
+      ),
+    }
+  } else {
+    return { value: response as string }
+  }
 }
 
 export const getServerSideProps: GetServerSideProps = async (ctx) => {