+++ packages/markdown/tsup.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'tsup'
 
 export default defineConfig({
   entry: ['src/index.tsx'],
-  clean: true,
+  clean: false,
   format: 'esm',
   dts: true,
 })